requirements.txt
google-cloud-pubsub==2.19.0
psutil==5.9.8
GPUtil==1.4.0
setuptools
